project/gmail_tools/src/requirements.txt
google-api-python-client
google-auth-oauthlib
requests
beautifulsoup4